Move CameraInfo to its own file / submodule
