+++ src/kind.rs
@@ -2,35 +2,17 @@
 
 use crate::common::*;
 
+mod camera_info;
 mod extension;
 mod frame_metadata;
 mod option;
 mod timestamp_domain;
 
+pub use camera_info::CameraInfo;
 pub use extension::Rs2Extension;
 pub use frame_metadata::Rs2FrameMetadata;
 pub use option::Rs2Option;
 pub use timestamp_domain::TimestampDomain;
-
-/// The enumeration of sensor information.
-#[repr(u32)]
-#[derive(FromPrimitive, Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum CameraInfo {
-    Name = sys::rs2_camera_info_RS2_CAMERA_INFO_NAME,
-    SerialNumber = sys::rs2_camera_info_RS2_CAMERA_INFO_SERIAL_NUMBER,
-    FirmwareVersion = sys::rs2_camera_info_RS2_CAMERA_INFO_FIRMWARE_VERSION,
-    RecommendedFirmwareVersion = sys::rs2_camera_info_RS2_CAMERA_INFO_RECOMMENDED_FIRMWARE_VERSION,
-    PhysicalPort = sys::rs2_camera_info_RS2_CAMERA_INFO_PHYSICAL_PORT,
-    DebugOpCode = sys::rs2_camera_info_RS2_CAMERA_INFO_DEBUG_OP_CODE,
-    AdvancedMode = sys::rs2_camera_info_RS2_CAMERA_INFO_ADVANCED_MODE,
-    ProductId = sys::rs2_camera_info_RS2_CAMERA_INFO_PRODUCT_ID,
-    CameraLocked = sys::rs2_camera_info_RS2_CAMERA_INFO_CAMERA_LOCKED,
-    UsbTypeDescriptor = sys::rs2_camera_info_RS2_CAMERA_INFO_USB_TYPE_DESCRIPTOR,
-    ProductLine = sys::rs2_camera_info_RS2_CAMERA_INFO_PRODUCT_LINE,
-    AsicSerialNumber = sys::rs2_camera_info_RS2_CAMERA_INFO_ASIC_SERIAL_NUMBER,
-    FirmwareUpdateId = sys::rs2_camera_info_RS2_CAMERA_INFO_FIRMWARE_UPDATE_ID,
-    Count = sys::rs2_camera_info_RS2_CAMERA_INFO_COUNT,
-}
 
 /// The enumeration of all categories of stream.
 #[repr(u32)]
