+++ src/kind/camera_info.rs
@@ -1,0 +1,27 @@
+//! The enumeration of sensor information.
+
+use num_derive::{FromPrimitive, ToPrimitive};
+use realsense_sys as sys;
+
+#[repr(u32)]
+#[derive(FromPrimitive, ToPrimitive, Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum CameraInfo {
+    Name = sys::rs2_camera_info_RS2_CAMERA_INFO_NAME,
+    SerialNumber = sys::rs2_camera_info_RS2_CAMERA_INFO_SERIAL_NUMBER,
+    FirmwareVersion = sys::rs2_camera_info_RS2_CAMERA_INFO_FIRMWARE_VERSION,
+    RecommendedFirmwareVersion = sys::rs2_camera_info_RS2_CAMERA_INFO_RECOMMENDED_FIRMWARE_VERSION,
+    PhysicalPort = sys::rs2_camera_info_RS2_CAMERA_INFO_PHYSICAL_PORT,
+    DebugOpCode = sys::rs2_camera_info_RS2_CAMERA_INFO_DEBUG_OP_CODE,
+    AdvancedMode = sys::rs2_camera_info_RS2_CAMERA_INFO_ADVANCED_MODE,
+    ProductId = sys::rs2_camera_info_RS2_CAMERA_INFO_PRODUCT_ID,
+    CameraLocked = sys::rs2_camera_info_RS2_CAMERA_INFO_CAMERA_LOCKED,
+    UsbTypeDescriptor = sys::rs2_camera_info_RS2_CAMERA_INFO_USB_TYPE_DESCRIPTOR,
+    ProductLine = sys::rs2_camera_info_RS2_CAMERA_INFO_PRODUCT_LINE,
+    AsicSerialNumber = sys::rs2_camera_info_RS2_CAMERA_INFO_ASIC_SERIAL_NUMBER,
+    FirmwareUpdateId = sys::rs2_camera_info_RS2_CAMERA_INFO_FIRMWARE_UPDATE_ID,
+    // Not included since this just tells us the total number of camera info options
+    //
+    // Count = sys::rs2_camera_info_RS2_CAMERA_INFO_COUNT,
+}
+
+
